+++ Final_Project_Conference_Room_Booking/Services/Implementation/ReservationHolderService.cs
@@ -26,23 +26,74 @@ namespace Final_Project_Conference_Room_Booking.Services.Implementation
         
         public async Task<ReservationHolder> Create(ReservationHolder reservationHolder)
         {
+            
+            if (reservationHolder == null)
+            {
+                throw new ArgumentNullException(nameof(reservationHolder), "The reservation holder cannot be null.");
+            }
+
+            
+            if (string.IsNullOrEmpty(reservationHolder.Name))
+            {
+                throw new ArgumentException("The Name field is required.", nameof(reservationHolder.Name));
+            }
+            if (string.IsNullOrEmpty(reservationHolder.Surname))
+            {
+                throw new ArgumentException("The Surname field is required.", nameof(reservationHolder.Surname));
+            }
+
+
+            if (string.IsNullOrEmpty(reservationHolder.Email))
+            {
+                throw new ArgumentException("The Email field is required.", nameof(reservationHolder.Email));
+            }
+       
+          
+            if (string.IsNullOrEmpty(reservationHolder.PhoneNumber))
+            {
+                throw new ArgumentException("The phone number field is required.", nameof(reservationHolder.PhoneNumber));
+            }
+
             return await _reservationHolderRepository.Create(reservationHolder);
         }
         public async Task<ReservationHolder> DeleteReservationHolder(int id)
         {
+            var reservationHolder = await _reservationHolderRepository.FindReservationHolder(id);
+            if (reservationHolder == null)
+            {
+                throw new ArgumentException("The reservation holder could not be found.", nameof(id));
+            }
+
+           
             return await _reservationHolderRepository.DeleteReservationHolder(id);
         }
         public async Task<ReservationHolder> FindReservationHolder(int id)
         {
-            return await _reservationHolderRepository.FindReservationHolder(id);
+            var reservationHolder = await _reservationHolderRepository.FindReservationHolder(id);
+            if (reservationHolder == null)
+            {
+                throw new ArgumentException("The reservation holder could not be found.", nameof(id));
+            }
+
+            return reservationHolder; ;
                 }
         public async Task<ReservationHolder> Edit(int id)
         {
-            return await _reservationHolderRepository.FindReservationHolder(id);
+            var reservationHolder = await _reservationHolderRepository.FindReservationHolder(id);
+            if (reservationHolder == null)
+            {
+                throw new ArgumentException("The reservation holder could not be found.", nameof(id));
+            }
+
+            return reservationHolder;
         }
         public async Task<ReservationHolder> Edit(ReservationHolder reservation)
         {
-            return await _reservationHolderRepository.Edit(reservation);
+            if (reservation == null)
+            {
+                throw new ArgumentException("There is no reservation");
+            }
+                return await _reservationHolderRepository.Edit(reservation);
         }
     }
 }
